refactor : post 관련 api 하나의 파일로 이동 & index.ts 정의
