+++ src/app/post/mentoring/[id]/layout.tsx
@@ -1,5 +1,5 @@
 import { queryclient } from "@/lib/getQueryClient";
-import getMentoringPost from "@/service/api/mentoring/post/getMentoringPost";
+import getMentoringPost from "@/service/api/post/getMentoringPost";
 import { dehydrate, HydrationBoundary } from "@tanstack/react-query";
 
 interface MentoringPostLayoutProps {
