+++ src/service/api/post/getAllMentoringPosts.ts
@@ -1,5 +1,5 @@
 import MentoringPosts from "@/types/post/mentoring/mentoringPosts";
-import { client } from "../../instance/client/client";
+import { client } from "../instance/client";
 
 export interface getAllMentoringPostsProps {
   nextCursor?: number;
@@ -10,9 +10,6 @@ export const getAllMentoringPosts = async ({
 }: getAllMentoringPostsProps): Promise<MentoringPosts> => {
   const response = await client.get(
     nextCursor ? `/mentoring/posts?cursor=${nextCursor}` : "/mentoring/posts"
-    // nextCursor
-    //   ? `mock/mentoring/posts?cursor=${nextCursor}`
-    //   : "mock/mentoring/posts"
   );
 
   return response.data.data;
