+++ src/service/api/post/index.ts
@@ -1,0 +1,10 @@
+import { getAllMentoringPosts } from "./getAllMentoringPosts";
+import { getAllProjectPosts } from "./getAllProjectPosts";
+import getMentoringPost from "./getMentoringPost";
+import getProjectPost from "./getProjectPost";
+
+// 프로젝트 포스트
+export { getProjectPost, getAllProjectPosts };
+
+// 멘토링 포스트
+export { getMentoringPost, getAllMentoringPosts };
